refactor: 유저페이지와 공통 컴포넌트 사용을 위한 폴더 구조 변경
