+++ src/app/(userpage)/_hooks/queries/usePurchaseList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { purchaseList } from "../../_api/mypage";
-import { calculateCounts } from "./useUserInfo";
+import { calculateCounts } from "../../_utils/calculateCounts";
 
 const usePurchaseList = () => {
   return useQuery({
