+++ src/app/(userpage)/_hooks/queries/useSaleList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { saleList } from "../../_api/mypage";
-import { calculateCounts } from "./useUserInfo";
+import { calculateCounts } from "../../_utils/calculateCounts";
 
 const useSaleList = (userId: number) => {
   return useQuery({
